+++ internal/app/app.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 
 	"github.com/chain4travel/camino-messenger-bot/config"
+	"github.com/chain4travel/camino-messenger-bot/events"
 	"github.com/chain4travel/camino-messenger-bot/internal/evm"
-	"github.com/chain4travel/camino-messenger-bot/internal/evm/events"
 	"github.com/chain4travel/camino-messenger-bot/internal/matrix"
 	"github.com/chain4travel/camino-messenger-bot/internal/messaging"
 	"github.com/chain4travel/camino-messenger-bot/internal/rpc/client"
